+++ nuxeo-drive-client/nxdrive/engine/workers.py
@@ -227,7 +227,7 @@ class PollWorker(Worker):
         self._check_interval = check_interval
         # Check at start
         self._next_check = 0
-        self.enable = True
+        self._enable = True
         self._metrics = {'last_poll': 0}
 
     def get_metrics(self):
@@ -249,7 +249,7 @@ class PollWorker(Worker):
         self._next_check = 0
 
     def _execute(self):
-        while self.enable:
+        while self._enable:
             self._interact()
             if self.get_next_poll() <= 0:
                 if self._poll():